+++ src/components/dialogs/index.ts
@@ -1,11 +1,13 @@
 import Dialog from './Dialog';
 import ConfirmDialog from './ConfirmDialog';
 import ConsentDialog from './ConsentDialog';
+import ReportDetailDialog from './ReportDetailDialog';
 import { defaultDialogPaddingXClass } from './constants';
 
 export {
   Dialog,
   ConfirmDialog,
   ConsentDialog,
+  ReportDetailDialog,
   defaultDialogPaddingXClass,
 }